unfold: Move Colors.h into Support directory as it is independent from BC.
